+++ app/src/main/java/com/example/arhiking/fragments/MapFragment.java
@@ -1,4 +1,4 @@
-package com.example.arhiking.ui.home;
+package com.example.arhiking.fragments;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -10,22 +10,23 @@ import androidx.annotation.NonNull;
 import androidx.fragment.app.Fragment;
 import androidx.lifecycle.ViewModelProvider;
 
-import com.example.arhiking.databinding.FragmentHomeBinding;
+import com.example.arhiking.databinding.FragmentMapBinding;
+import com.example.arhiking.viewmodels.MapViewModel;
 
-public class HomeFragment extends Fragment {
+public class MapFragment extends Fragment {
 
-    private FragmentHomeBinding binding;
+    private FragmentMapBinding binding;
 
     public View onCreateView(@NonNull LayoutInflater inflater,
                              ViewGroup container, Bundle savedInstanceState) {
-        HomeViewModel homeViewModel =
-                new ViewModelProvider(this).get(HomeViewModel.class);
+        MapViewModel mapViewModel =
+                new ViewModelProvider(this).get(MapViewModel.class);
 
-        binding = FragmentHomeBinding.inflate(inflater, container, false);
+        binding = FragmentMapBinding.inflate(inflater, container, false);
         View root = binding.getRoot();
 
-        final TextView textView = binding.textHome;
-        homeViewModel.getText().observe(getViewLifecycleOwner(), textView::setText);
+        final TextView textView = binding.textMap;
+        mapViewModel.getText().observe(getViewLifecycleOwner(), textView::setText);
         return root;
     }
