+++ app/src/main/java/com/example/arhiking/fragments/SettingsFragment.java
@@ -1,4 +1,4 @@
-package com.example.arhiking.ui;
+package com.example.arhiking.fragments;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -10,22 +10,22 @@ import androidx.annotation.NonNull;
 import androidx.fragment.app.Fragment;
 import androidx.lifecycle.ViewModelProvider;
 
-import com.example.arhiking.databinding.FragmentMapBinding;
-import com.example.arhiking.databinding.FragmentRegisterHikeBinding;
+import com.example.arhiking.databinding.FragmentSettingsBinding;
+import com.example.arhiking.viewmodels.SettingsViewModel;
 
-public class RegisterHikeFragment extends Fragment {
+public class SettingsFragment extends Fragment {
 
-    private FragmentRegisterHikeBinding binding;
+    private FragmentSettingsBinding binding;
 
     public View onCreateView(@NonNull LayoutInflater inflater,
                              ViewGroup container, Bundle savedInstanceState) {
-        RegisterHikeViewModel mapViewModel =
-                new ViewModelProvider(this).get(RegisterHikeViewModel.class);
+        SettingsViewModel mapViewModel =
+                new ViewModelProvider(this).get(SettingsViewModel.class);
 
-        binding = FragmentRegisterHikeBinding.inflate(inflater, container, false);
+        binding = FragmentSettingsBinding.inflate(inflater, container, false);
         View root = binding.getRoot();
 
-        final TextView textView = binding.textRegisterHike;
+        final TextView textView = binding.textSettings;
         mapViewModel.getText().observe(getViewLifecycleOwner(), textView::setText);
         return root;
     }
