+++ app/src/main/java/com/example/arhiking/viewmodels/HomeViewModel.java
@@ -1,16 +1,16 @@
-package com.example.arhiking.ui;
+package com.example.arhiking.viewmodels;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-public class UserViewModel extends ViewModel {
+public class HomeViewModel extends ViewModel {
 
     private final MutableLiveData<String> mText;
 
-    public UserViewModel() {
+    public HomeViewModel() {
         mText = new MutableLiveData<>();
-        mText.setValue("This is user fragment");
+        mText.setValue("This is home fragment");
     }
 
     public LiveData<String> getText() {
