+++ app/src/main/java/com/example/arhiking/viewmodels/RegisterHikeViewModel.java
@@ -1,16 +1,16 @@
-package com.example.arhiking.ui.home;
+package com.example.arhiking.viewmodels;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-public class HomeViewModel extends ViewModel {
+public class RegisterHikeViewModel extends ViewModel {
 
     private final MutableLiveData<String> mText;
 
-    public HomeViewModel() {
+    public RegisterHikeViewModel() {
         mText = new MutableLiveData<>();
-        mText.setValue("This is home fragment");
+        mText.setValue("This is register hike fragment");
     }
 
     public LiveData<String> getText() {
