+++ app/src/main/java/com/example/arhiking/viewmodels/SettingsViewModel.java
@@ -1,16 +1,16 @@
-package com.example.arhiking.ui;
+package com.example.arhiking.viewmodels;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-public class RegisterHikeViewModel extends ViewModel {
+public class SettingsViewModel extends ViewModel {
 
     private final MutableLiveData<String> mText;
 
-    public RegisterHikeViewModel() {
+    public SettingsViewModel() {
         mText = new MutableLiveData<>();
-        mText.setValue("This is register hike fragment");
+        mText.setValue("This is settings fragment");
     }
 
     public LiveData<String> getText() {
