+++ app/src/main/res/layout/fragment_register_hike.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.RegisterHikeFragment">
+    tools:context=".fragments.RegisterHikeFragment">
 
     <TextView
         android:id="@+id/text_register_hike"
