+++ app/src/main/res/layout/fragment_user.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.UserFragment">
+    tools:context=".fragments.UserFragment">
 
     <TextView
         android:id="@+id/text_user"
