+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -7,31 +7,31 @@
 
     <fragment
         android:id="@+id/navigation_home"
-        android:name="com.example.arhiking.ui.home.HomeFragment"
+        android:name="com.example.arhiking.fragments.HomeFragment"
         android:label="@string/title_home"
         tools:layout="@layout/fragment_home" />
 
     <fragment
         android:id="@+id/navigation_map"
-        android:name="com.example.arhiking.ui.dashboard.MapFragment"
+        android:name="com.example.arhiking.fragments.MapFragment"
         android:label="@string/title_map"
         tools:layout="@layout/fragment_map" />
 
     <fragment
         android:id="@+id/navigation_user"
-        android:name="com.example.arhiking.ui.UserFragment"
+        android:name="com.example.arhiking.fragments.UserFragment"
         android:label="@string/title_user"
         tools:layout="@layout/fragment_user" />
 
     <fragment
         android:id="@+id/navigation_register_hike"
-        android:name="com.example.arhiking.ui.RegisterHikeFragment"
+        android:name="com.example.arhiking.fragments.RegisterHikeFragment"
         android:label="@string/title_register_hike"
         tools:layout="@layout/fragment_register_hike" />
 
     <fragment
         android:id="@+id/navigation_settings"
-        android:name="com.example.arhiking.ui.SettingsFragment"
+        android:name="com.example.arhiking.fragments.SettingsFragment"
         android:label="@string/title_settings"
         tools:layout="@layout/fragment_settings" />
 </navigation>
